// ui/src/app/footprint/generated/footprint_completed_integration_event.ts
export type T = {
  instrument: InstrumentViewModel;
  timeframe: string;
  open_ts: string;
  open_price: string;
  high: string;
  low: string;
  close: string;
  volume: string;
  delta: string;
  poc_price: string;
  clusters: ClusterViewModel[];
}